@@ -1,9 +1,16 @@
 module github.com/subpop/mqttcli
 
-go 1.16
+go 1.20
 
 require (
 	github.com/eclipse/paho.mqtt.golang v1.4.3
 	github.com/peterbourgon/ff/v3 v3.4.0
 	github.com/sgreben/flagvar v1.10.1
+)
+
+require (
+	github.com/gobwas/glob v0.2.3 // indirect
+	github.com/gorilla/websocket v1.5.0 // indirect
+	golang.org/x/net v0.8.0 // indirect
+	golang.org/x/sync v0.1.0 // indirect
 )
